# bike-stores-client/src/app/libs/api/v1/model/category.ts
/**
 * bike-stores-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Product } from './product';


export interface Category { 
    categoryId?: number;
    categoryName?: string | null;
    products?: Array<Product> | null;
}