# bike-stores-client/src/app/libs/api/v1/model/order-item.ts
/**
 * bike-stores-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Order } from './order';
import { Product } from './product';


export interface OrderItem { 
    orderId?: number;
    itemId?: number;
    productId?: number;
    quantity?: number;
    listPrice?: number;
    discount?: number;
    order?: Order;
    product?: Product;
}