# bike-stores-client/src/app/libs/api/v1/model/product.ts
/**
 * bike-stores-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Brand } from './brand';
import { Category } from './category';
import { OrderItem } from './order-item';
import { Stock } from './stock';


export interface Product { 
    productId?: number;
    productName?: string | null;
    brandId?: number;
    categoryId?: number;
    modelYear?: number;
    listPrice?: number;
    brand?: Brand;
    category?: Category;
    orderItems?: Array<OrderItem> | null;
    stocks?: Array<Stock> | null;
}