# bike-stores-client/src/app/libs/api/v1/model/stock.ts
/**
 * bike-stores-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Product } from './product';
import { Store } from './store';


export interface Stock { 
    storeId?: number;
    productId?: number;
    quantity?: number | null;
    product?: Product;
    store?: Store;
}