# bike-stores-client/src/app/libs/api/v1/model/store.ts
/**
 * bike-stores-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Order } from './order';
import { Staff } from './staff';
import { Stock } from './stock';


export interface Store { 
    storeId?: number;
    storeName?: string | null;
    phone?: string | null;
    email?: string | null;
    street?: string | null;
    city?: string | null;
    state?: string | null;
    zipCode?: string | null;
    orders?: Array<Order> | null;
    stocks?: Array<Stock> | null;
    staff?: Array<Staff> | null;
}